refactor: split monolithic commands.rs into per-domain modules (#151) (#185)
Break the 3115-line commands.rs into 11 focused modules under commands/:
workspace, ship, pr, ci, diff, history, stack, tracker_cmds, config,
doctor, release. No behavioral changes — pure structural refactor.
