+++ src/cli/commands/mod.rs
@@ -1,0 +1,23 @@
+mod ci;
+mod config;
+mod diff;
+mod doctor;
+mod history;
+mod pr;
+mod release;
+mod ship;
+mod stack;
+mod tracker_cmds;
+mod workspace;
+
+pub use ci::*;
+pub use config::*;
+pub use diff::*;
+pub use doctor::*;
+pub use history::*;
+pub use pr::*;
+pub use release::*;
+pub use ship::*;
+pub use stack::*;
+pub use tracker_cmds::*;
+pub use workspace::*;
